Catch shareId from upload share mutation
